requirements.txt
numpy==1.24.4
opencv-python==4.8.0
streamlit==1.27.0
Pillow==8.2.0
```